src/pyproject.toml
[tool.poetry]
name = "ai-pair-programmer"
version = "0.1.0"
description = "AI pair programmer tool"
authors = ["Ronnasayd Machado"]
packages = [{ include = "mcps" }]

[tool.poetry.dependencies]
python = "^3.11"
fastmcp = "^2.0"
pymongo = "==4.16.0"

# [tool.poetry.group.dev.dependencies]
# pytest = "^7.0"
# pytest-asyncio = "^0.23.0"
# pytest-mock = "^3.12.0"

[tool.poetry.scripts]
mongodb-mcp = "mcps.mongodb_mcp:main"
mysql-mcp = "mcps.mysql_mcp:main"
postgresql-mcp = "mcps.postgresql_mcp:main"
sqlite-mcp = "mcps.sqlite_mcp:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
